Remove unnecessary sleep in ActivityThread.
This attachApplication function is working when ActivityThread is created and it takes 30~40 milliseconds in G3.
At that time, ActivityThread goes to sleep even it is not necessary.
Parallel processing can be helpful in this case.
After merging this code, there is no sleep time during attachApplication function and
starting application time is reduced.

sleep time 30ms -> 0ms in G3.

Change-Id: Ia4babb97d15234ac430d148cfdcadfbc7689bb74
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>